requirements.txt
google-search-results>=2.4.2
python-dotenv>=1.0.0
requests>=2.31.0
